Reorganized sceneView and receiver
